Use a pattern synonym to unpack Refl in Sphere.
While we weren't able to unpack Refl in the last step because it
is a sum type, we can if we use a trick modeled after an C Enum.
The use of COMPLETE is unsafe here because other values could be
constructed. For this small example though we rely on the fact
that we create values of Refl via the patterns.