Remove mutability.
By removing mutability we remove a lot of barrier to optimization.
There is the potential for fusion that was formerly blocked by
references and the mutable vector, in addition to the passes on
y being independent if we ever wanted to parallelize procesing.

Reordering the order we do 'y's in allows us to not use an
intermediary structure at all and directly generate  the result.
Since we're generating results directly, we just produce a list,
and don't bother with the 'i' variable for placement, which means
we want to reorder which y rows we generate first. So we generate
them in the reverse order now (but this does potentially change
edge cases and parameter validation should be added to avoid any
issues with degenerate perameters).