Strategic application of strictness.
Here we go through the program and apply some considered strictness
to it. Forcing computation we don't need to do, or often forcing it
to far before we need it is a loss. When we put a bang pattern in,
GHC doesn't move it so we have to hand float it to an appropriate
location.