Remove Maybe from intersect(s).
Our innermost functions are of critical importance. Here we remove
a Maybe which significantly reduces the boxing (which could have
been mitigated with a StrictMaybe) and the cases.

Since a Ray that fails to intersect something can be said to
intersect at infinity, Double already actually covers the
structure at play.

This also reduces allocation.