Custom datatype for 'intersects' parameter passing.
Its clear that f will consume these, so making it strict can have
some benefit. The problem is the right level here. One's first
take might be to set the parameters of 'f' to be strict, and that
can have benefits, but it leaves a lot on the table. Moving to
and unboxed tuple performs the computation faster but introduces
copying. Out 'T' datatype here has the right balance, it makes it
clear the data is strict but doesn't copy.