Fix difference with C++ version.
Since the sha1 of the output didn't match the C++ version we
started investigating. We noticed clang++ and g++ actually
produced different sha1s, but more importantly for us, we found
that the unincrimented depth was being used in one branch. Fixing
that, we can generate an image with an identical sha1 to the
clang++ version so we are confident in saying that we are doing
the same thing in the same way as the C++ code.

Generating equal sha1s is sensative to numerical inaccuracy and
generating identical sha1s is more reliable < 150x150. A ray-by-ray
analysis shows some rays can differ in the last decimal digit of the
Double, and we conjecture that this same floating point instability
accounts for the g++/clang++ hash difference.